Provide Section-Specific `BYTECODE`
This updates the code so that it provides the actual code being executed
for each section.  However, its not really clear to me that this
matters.  In particular, only legacy contracts need this information
(i.e. to determine JUMPDEST).

The outstanding problem at this stage is that we have lost stack
information on entry to each block.