Refactor block printer
Mechanism for printing chunks of blocks into separate files (aka
groups) is now in place.  However, at this stage, the algorithm for
splitting up the initial set of blocks is essentially missing.

Therefore, the next steps are: (1) add support for splitting blocks into
groups; (2) enable configuration of the split via the cli.