Dominator analysis plumbed in.
The dominator analysis is now (hopefully) working correctly.  What
remains is to actually read in a root configuration, and act
accordingly.  Furthermore, there remain outstanding issues for the
"bottom" part of the cfg.  That is, stuff which is not owned by any root
but is reachable from a root.  The problem is that this is currently
folded into the topmost root.