Code Splitting Activated
The algorithm for splitting a CFG into separate files is now
operational, though a number of features remain to be implemented.
Specifically, outstanding issues are:

(1) Missing includes.  Currently, appropriate includes are not added.

(2) Missing "bottom".  Currently, everthing that is not given in a root
is retained in the topmost file.  That clearly won't work!