Big refactoring
This now refactors the minimisation algorithm to integrate it more
closely with the concept of a `Block`.  In particular, we now retain
information at each program point which can be useful for debugging.

I think we need some command-line switches to signal how and when to
print the debug minisation info, and whether or not intermediate
comments should be expanded or not.